fix: 지하철 소요 시간 Type 수정

+++ src/models/subway.py
@@ -21,4 +21,4 @@ class SubwayRouteStation(BaseModel):
     route_id: Mapped[int] = mapped_column(nullable=False)
     station_name: Mapped[str] = mapped_column(String(30), nullable=False)
     station_sequence: Mapped[int] = mapped_column(nullable=False)
-    cumulative_time: Mapped[int] = mapped_column(nullable=False)
+    cumulative_time: Mapped[float] = mapped_column(nullable=False)
